Fix whitespace issues and other minor typos.
Removed redundant span tags that were causing spaces before commas and
periods. Fixed other minor typos. Changed title on first slide to title
case in the HTML instead of the inconsistent case it was in (it is
displayed as ALL CAPS either way, so there is no viewer visible change
from this particular alteration).